refactor int to ong long

+++ include/BigNum.h
@@ -8,6 +8,6 @@ void formatStr(std::string& first, std::string& second);
 std::string addStr(std::string first, std::string second);
 char multChar(char first, char second, int& inc);
 std::string multStrAndChar(std::string first, char second);
-std::string toStr(int num);
+std::string toStr(long long num);
 
 #endif
